+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 # @project-version@ (@build-time@)
 
-* Fixed a Lua error that occurred on Death Knights when Blizzard-style combo points were used in Mists Classic.
+* Added an option to change the healthbar border color [GH-600 by PolyBadger].